refactor : InputTest 클래스를 세분화
